requirements.txt
PyMuPDF
customtkinter
pyinstaller
Pillow
